Refactored event handler standalone module, replaced epoch/step references with Timestamp object
