+++ composer/profiler/__init__.py
@@ -1,6 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
+from composer.profiler.event_handler import ProfilerEventHandler as ProfilerEventHandler
 from composer.profiler.profiler import Profiler as Profiler
 from composer.profiler.profiler import ProfilerAction as ProfilerAction
-from composer.profiler.profiler import ProfilerEventHandler as ProfilerEventHandler
 from composer.profiler.profiler_hparams import ProfilerHparams as ProfilerHparams
